requirements.txt
Django==1.10.1
Pillow
MySQL-python
django-crispy-forms==1.6.0
django-registration-redux==1.4
python-social-auth==0.3.3